+++ test/chez/scan-error-app/src/app/core.clj
@@ -1,0 +1,3 @@
+(ns app.core
+  (:require [app.dup]))
+(defn -main [& _] (println (app.dup/f)))